[ty] Refactor with find references https://github.com/astral-sh/ty/issues/3514
