+++ crates/ty_ide/src/lib.rs
@@ -18,11 +18,13 @@ mod goto_implementation;
 mod goto_type_definition;
 mod hints;
 mod hover;
+mod implementations;
 mod importer;
 mod inlay_hints;
 mod markup;
 mod references;
 mod rename;
+mod scan;
 mod selection_range;
 mod semantic_tokens;
 mod signature_help;
